requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0

# Database
asyncpg==0.29.0
redis==5.0.1
psycopg2-binary==2.9.9

# Vector Database
pymilvus==2.3.3

# Image processing and recognition
face-recognition==1.3.0
face-recognition-models==0.3.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# OCR for ID verification
pytesseract==0.3.10

# LLM Support
llama-cpp-python==0.2.20
huggingface-hub==0.19.4

# LangChain and RAG
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.7
langchain-experimental==0.0.47
chromadb==0.4.20
faiss-cpu==1.7.4
tiktoken==0.5.2

# NLP and ML
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
tokenizers==0.15.0

# Data processing
pandas==2.1.3
openpyxl==3.1.2
scikit-learn==1.3.2
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
setuptools<81  # Pin to avoid pkg_resources deprecation

# Performance
aiofiles==23.2.1
httpx==0.25.1
asyncio==3.4.3

# Monitoring (optional)
prometheus-client==0.19.0

# Graph Database
neo4j==5.15.0
py2neo==2021.2.3

# Semantic Search
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0

# Graph Analytics
networkx==3.2.1
python-louvain==0.16
scipy==1.11.4

# Natural Language Processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0